Using autoconf to figure out 64-bit types. Avoid off_t and 64-bit constants.
darcs-hash:20050908152635-65a35-f0bb1705406948290003acb83f099c0be9167bda.gz

+++ libsofia-sip-ua/nta/sl_read_payload.c
@@ -79,7 +79,7 @@ sip_payload_t *sl_fread_payload(su_home_t *home, FILE *f)
   int n;
   char *buf;
   char const *who;
-  off_t used, size;
+  size_t used, size;
 
   if (f == NULL) {
     errno = EINVAL;
